Conflicts:
	build.number
	build.xml
	commons/pom.xml
	core/src/main/java/com/orientechnologies/orient/core/config/OGlobalConfiguration.java
	core/src/main/java/com/orientechnologies/orient/core/index/OPropertyListIndexDefinition.java
	core/src/main/java/com/orientechnologies/orient/core/metadata/schema/OClass.java
	core/src/main/java/com/orientechnologies/orient/core/sql/filter/OSQLFilterItemField.java
	core/src/main/java/com/orientechnologies/orient/core/sql/functions/OSQLFunctionRuntime.java
	core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/OStorageLocal.java
	core/src/main/java/com/orientechnologies/orient/core/type/tree/OMVRBTreePersistent.java
	distributed/config/default-distributed-db.json
	distributed/config/orientdb-dserver-config.xml
	distributed/pom.xml
	distributed/src/test/resources/default-distributed-db.json
	distributed/src/test/resources/orientdb-dserver-config-1.xml
	distributed/src/test/resources/orientdb-dserver-config-2.xml
	distributed/src/test/resources/orientdb-dserver-config-3.xml
	distribution/pom.xml
	graphdb/pom.xml
	object/src/main/java/com/orientechnologies/orient/object/db/OObjectDatabaseTx.java
	object/src/main/java/com/orientechnologies/orient/object/enhancement/OObjectEntitySerializer.java
	object/src/main/java/com/orientechnologies/orient/object/enhancement/OObjectProxyMethodHandler.java
	object/src/main/java/com/orientechnologies/orient/object/enumerations/OObjectEnumLazyList.java
	object/src/main/java/com/orientechnologies/orient/object/enumerations/OObjectEnumLazySet.java
	pom.xml
	server/src/main/java/com/orientechnologies/orient/server/network/protocol/binary/OBinaryNetworkProtocolAbstract.java
	server/src/main/java/com/orientechnologies/orient/server/network/protocol/binary/ONetworkProtocolBinary.java
	tests/pom.xml
	tests/src/test/java/com/orientechnologies/orient/test/database/auto/ObjectDetachingTest.java

+++ server/src/main/java/com/orientechnologies/orient/server/task/ODeleteRecordDistributedTask.java
@@ -25,6 +25,7 @@ import com.orientechnologies.orient.core.id.ORecordId;
 import com.orientechnologies.orient.core.record.ORecordInternal;
 import com.orientechnologies.orient.server.distributed.ODistributedServerManager.EXECUTION_MODE;
 import com.orientechnologies.orient.server.distributed.OStorageSynchronizer;
+import com.orientechnologies.orient.server.distributed.conflict.OReplicationConflictResolver;
 import com.orientechnologies.orient.server.journal.ODatabaseJournal.OPERATION_TYPES;
 
 /**
@@ -52,16 +53,32 @@ public class ODeleteRecordDistributedTask extends OAbstractRecordDistributedTask
   protected Boolean executeOnLocalNode(final OStorageSynchronizer dbSynchronizer) {
     OLogManager.instance().warn(this, "DISTRIBUTED <-[%s/%s] DELETE RECORD %s v.%d", nodeSource, databaseName, rid, version);
 
-    final ODatabaseDocumentTx database = getDatabase();
+    final ODatabaseDocumentTx database = openDatabase();
     try {
       final ORecordInternal<?> record = database.load(rid);
-      record.setVersion(version);
-      record.delete();
-
-      return Boolean.TRUE;
+      if (record != null) {
+        record.setVersion(version);
+        record.delete();
+        return Boolean.TRUE;
+      }
+      return Boolean.FALSE;
     } finally {
-      database.close();
+      closeDatabase(database);
     }
+  }
+
+  /**
+   * Handles conflict between local and remote execution results.
+   * 
+   * @param localResult
+   *          The result on local node
+   * @param remoteResult
+   *          the result on remote node
+   */
+  @Override
+  public void handleConflict(final String iRemoteNodeId, final Object localResult, final Object remoteResult) {
+    final OReplicationConflictResolver resolver = getDatabaseSynchronizer().getConflictResolver();
+    resolver.handleDeleteConflict(iRemoteNodeId, rid);
   }
 
   @Override
